Importing bootstrap website I designed on separate repository.
This commit is the initial website commit being made directly to the master. It directly replaces Index.html, but otherwise shouldn't muck about with much else existing. May need a little bit of cleanup, but should mostly work fine. Output box in tournament.html is a trial tool and has not been tested.